[IMP] helpdesk: review the ratings in helpdesk
Before this commit, we compute the customer satisfaction only for the
closed tickets with a closed date greater or equal to today (for happy
ratings for today) or a closed date greater or equal to the date 7 days
before (for the happy ratings from 1 week in the past). We also compute
the customer_satisfaction field in the helpdesk.team but this field is
in fact the rating average. Since we add rating_avg field in
rating_mixin, we can replace the `customer_satisfaction` field by
`rating_avg` field.

This commit replaces the `customer_satisfaction` by the `rating_avg` and
display the emoji in the kanban view of helpdesk team as we done for the
one of helpdesk ticket. Moreover, the ratings in the helpdesk dashboard
is review to compute the rating average of today of tickets belonging to
the current user and also the rating average from 7 days before until
today for the tickets belonging to the current user.

task-2671848

Community PR: odoo/odoo#81027

Closes #22773