[FIX] helpdesk: check user can access to team with usa_sla or use_rating
Before this commit, when the current user accesses to the helpdesk app,
we compute his stats to display in the dashboard. For this stat, we
check if the user can see at least a team with those features enabled
and he is a member of the team. The problem is to check if the user is a
member of the team we use the `member_ids` field and this field is only
used for the random and balanced assignment methods. So, for the
helpdesk team in which the assignment method is always the manual one
than the `member_ids` will contain by default the user who created the
helpdesk team and never changed if the assignment method never changes
too. Based on this, if another user than the one whose create the
helpdesk teams, then this user could see no stat for SLA and Ratings
features since he will not be a member of a helpdesk team.

This commit changes domain to check if the user can see at least a team.
We remove the condition to check if he is a member of the team. Because,
we can just let the ir.rules work to filter and get the team.

task-2671848

Community PR: odoo/odoo#81027

Closes #22773

closes odoo/enterprise#22956

X-original-commit: 4d09ec45396362224cc5ca4becf5f90ff81a8259
Related: odoo/odoo#81553
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>