[FIX] {sale_,}planning: fix access rights on copy and plan
Purpose
=======
An issue happen when users with low-level access opens the planning app,
uses the copy previous week or plan a slot on a database using
hr_contract module.
Those users don't have access to employee contracts.

The implementation now makes the hypothesis those functionalities should
always work with the same behavior whatever the user rights or groups
are, i.e. the resource intervals information should be used as a black
box.
NB : The user still needs to have the right access to planning, only the
methods to get the resource intervals are blackboxed/used with sudo.

Regarding the progress bar, this one should work for all planning users
whom have rights to read on the resources displayed in the gantt.

task-2705003

closes odoo/enterprise#22743

X-original-commit: 1a9cbad4606d3c6914fde9de5a503e6425f33ae8
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>