[FIX] l10n_mx_edi: prevent traceback when downloading a CFDI that fails XSD validation
`_get_invoice_edi_content` used by the download feature expects a CFDI XML string to be populated. When the CFDI document fails XSD validation - the `cfdi_str` value is not available in the dict returned by `_l10n_mx_edi_export_invoice_cfdi`
With this commit, we ensure that the cfdi_str is always returned along with the errors.

closes odoo/enterprise#22875

X-original-commit: a3c6d7ad5fbd2ab82cd3e6f77a1fc0ff28e9335a
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Ayob Habib (ayh) <ayh@odoo.com>