[FIX] web_studio: Fix form_editor editing with studio
In the case where a form view contains enough element to be able to be scrolled, we cannot add a new field in the scrolled area
Because there is a css: Height: 100% property.

Because of this Height jquery-ui does not detect the new element (draggable) because it is not in the (droppable) area because it has been scrolled up.

this was originally from: https://github.com/odoo/enterprise/blob/aabb60e7a89e2d0a62471b397b434e43a205e2f9/web_enterprise/static/src/legacy/scss/form_view.scss#L332-L334

opw-2695114

closes odoo/enterprise#22860

X-original-commit: 04a0755a245ad5edc9153734fef6af630ae8cfe9
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>