[FIX] Differentiate virtual ids for DATEV between suppliers and customers
Before : all the partners were in the same file and with the same prefix, which is not accepted (the prefixes for suppliers and customers are not in the same range).

After : the partners are separated between suppliers and customers (different files and different prefixes).

Defined a second system parameter and used it for suppliers, while the previous one will be used for customers (without changing name of the first one)

Tests were minimal so added tests for the expected csv file for different types of entries, and for payments.

closes odoo/enterprise#24967

X-original-commit: acf134a244cd98d3bee29db4509e58192fcb5301
Signed-off-by: Olivier Colson <oco@odoo.com>