[IMP] appointment: transform answers into a O2M field.
Before, the answers were a many2many field on calendar.appointment.question
model. However, it did not make much sense to have all existing answers
available to choose from when selecting pre-existing answers.

After this commit, the field answer_ids is set as a one2many instead, only
containing meaningful answers. The use of a o2m is therefore natural. This
makes the many2many widget deficient in the question form view, leading to
its removal. A further commit should make it more user-friendly.

An exception is raised if no answer is set when using a selection mode
(select / radio / checkbox). We set copy=True on the answers in case of
a duplicate since they are constrained to exist and O2m now.

Task-2621327

Part-of: odoo/enterprise#20433