[IMP] appointment: add appointment question answer input model and reporting
In order for the appointment manager to see what the participants answer,
this commit adds reporting features for each question, directly accessible
via the bar chart icon on the question tree view in appointment type form.
- select/radio/checkbox question_type: pivot/graph/tree
- char/text question_type: tree
A form view is also added in order to see each answer input in detail and
properly structured when clicking on a line in the tree view.

To do so, a new model calendar.appointment.answer.input is added, and answers
are now added in database when the appointment form is submitted. Before, they
were only stored as a formatted text: the description of the calendar_event
created for the appointment.
- select/radio : store id of selected answer
- checkbox : create one record per selected answer and store it
- text/char : store the text answer

Also, the answers are linked to the calendar event created on submission.
Indeed, they are created directly when the event is created, through commands.
This makes the event 'aware' of the inputs written down on the appointment
form by the attendee (and vice versa). This is done adding a M2O / O2M
relation on models calendar_appointment_answer_input and calendar_event.

They are also linked to the customer taking the appointment, for reporting purpose.
The answer input search view receives therefore a few filters / group_bys
in order to give more accurate reporting instead of anonymous numbers, and the
tree view matches most new fields.

Task-2621327

Part-of: odoo/enterprise#20433