[REF] web_map: ease map view use in standalone
In the typical usecase of views (i.e. with the action service), a
config is defined in the env and contains a lot of information, for
instance related to control panel, breadcrumbs, view switcher...
However, views are designed to be used in standalone as well, in
which case there's no guarantee that information would be in the
config (which is empty by default).

Before this commit, the MapView relied on the presence of the
"views" key in the config, and it crashed if it wasn't set. This
commit adds a fallback to the empty array.

Part-of: odoo/enterprise#22799