[IMP] appointment: improve display of answers in back-end appointment views
The display of answer_ids is improved and switched to an editable tree view
which makes it simpler to edit, see and remove answers. A field sequence
is added to answers in order for them to be sorted with handle widget.

In the appointment_type form, the imbedded question tree view now displays
available answers (in order) with a many2many tags widget. The user do not
have to click on it and open the menu to see what available answers there
are.

Task-2621327

Part-of: odoo/enterprise#20433