[IMP] helpdesk: check if SLA or rating are enabled via hook
Before this commit, to check if the SLA feature is enabled for the
current user, we need to check if the user can see a helpdesk team with
`usa_sla` is true and he has the `group_use_sla` group. (Same process
for the customer ratings feature).

This commit adds 2 private methods used as a hook to check if the
feature is enabled for the current user. Thanks to this commit, we just
need to use this method to know if the feature is available instead of
writing the search method and the user_has_group method each time we
need to check if one of both features is enabled for the current user.

task-2671848

Community PR: odoo/odoo#81027

Closes #22773