[FIX] helpdesk: fix the unassigned ticket domain for team
Currently, when click on unassigned ticket from kanban card of
the helpdesk team it will display the tickets of all the teams
instead of particular team.

So in this commit, pass the domain for particular team so when
click on unassigned team from team kanban view it will display the
ticket of particular team.

Closes: https://github.com/odoo/enterprise/pull/22820
task-2712024

closes odoo/enterprise#23230

X-original-commit: 86529d46ab13dd75ac6d03d0fb916fd066c17b82
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>