[IMP] appointment: compare ids of question answers instead of names
Before, when submitting an appointment selection form, it returned
the names in case of 'select' and 'radio' fields, used to retreive
existing answers. This could lead to mismatches (answer name changed,
duplicates, ...) and therefore, they are switched to the ids directly.
This ensures adaptablility and is cleaner.

Task-2621327

Part-of: odoo/enterprise#20433