[IMP] delivery_fedex: rate v16->31 & ship v15->28
We were using an outdated version of FedEx Web Services for which support will
only be available to a very limited extent as of January 3, 2022 and we will no
longer be able to access their new services with the older version.

There are some additions with the new versions, but it looks totally compatible
with our implementation.

task-2357856

closes odoo/enterprise#22900

X-original-commit: b399aa1d3b0527c0b6c2694e8ddfb12adeca4a79
Signed-off-by: Arnold Moyaux <arm@odoo.com>