[FIX] account_asset: fix duration/rate in depreciation schedule report.
Before this fix, consider an asset (method linear) with the exisiting depreciation is 100% of depreciations (when importing data to odoo of a already 100% deprecated asset). The depreciation schedule report would show "0%" in the Rate (14.0) or Duration / Rate (15.0) column instead of the correct rate or the depreciation duration.
This came from an error in computation of the total duration (duration + depreciation_number_import) in the SQL request when duration is 0.

closes odoo/enterprise#22935

Task: 2702612
X-original-commit: 3faca2def4e4f481d8d1fa084e58a182cb6ad388
Signed-off-by: William André (wan) <wan@odoo.com>