[IMP] social: cleanup '_htmlEscape' usage
This commit cleans our custom implementation of HTML escaping and uses the
available tools from owl.utils.

This custom implementation was made to avoid escaping '&' because that would
break link parameters (e.g: 'https://odoo.com?a=1&b=2').
However, by slightly adapting our URL regex, we can account for this escaping
and properly extract URLs even if they are escaped.

This change comes with a test to make sure we still support URL parameters.

In addition, we slightly adapted the hashtag regex to ensure it's preceded by
the start of line or a blank space character.

Finally, we adapted some demo data in the social_demo module to ease comment
posting on Twitter on the related demo accounts.

Task-2622893

Part-of: odoo/enterprise#22850