[IMP] [website_]appointment[_hr]: reword staff_user_ids.
The staff_user_ids are the users of the staff on appointment types.

This commit keeps the 'staff_user' in code, but rewords docstrings
and litteral occurences of 'Staff member' to a simplier 'User'.
This reduces the feeling of a new concept introduced.

Staff User / Staff Member can feel like another model but is simply
users. Therefore, this reduces the confusion.

Task-2499566

closes odoo/enterprise#17934

Related: odoo/odoo#82234
Related: odoo/upgrade#2578
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>