[FIX] web_mobile: fix download report
The webview in Android don't support the blob download file.

In `web_mobile` an override was made to call a native method.
After to the WOWL refactoring, this override still exist but only
for the legacy part.

This commit adds the counter part for the new WebClient.

Steps to reproduce:
* Open Odoo in the Android Mobile App
* Go to 'Inventory' app
* Select 'Receipts'
* Select an receipt (e.g. WH/IN/00002)
* Open the 'Action' dropdown
* Select Print
* A toast is shown ("The Odoo Mobile Apps only supports file downloads
  using the HTTP protocol.") => BUG

opw-2702665

closes odoo/enterprise#22944

X-original-commit: a66d0f9566ad75defbea58d8c2203b160756866d
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>