[IMP] sale_timesheet_enterprise: recompute all SOL in batch when invoiced_timesheet changed
Before this commit, when the sale.invoiced_timesheet settings changes,
we search all Sales Orders with state is equal to 'sale' or 'done' and
then we search the SOL contained a service product, more precisely, a
timesheet product. And then we call the compute methods for all SOLs
found per SO

This commit directly search the SOLs based on the condition described
above and calls the compute methods for those SOLs. So, instead of
calling compute methods per SO, we call once if we found at least an SOL
to recompute.

task-2648955

Community PR: odoo/odoo#82379

Closes #23315