[IMP] sale_timesheet_enterprise: filter timesheets in project sharing
Before this commit, when the portal user goes to a task via the project
sharing feature and the `sale.invoiced_timesheet` is equal to 'approved'
then he can only see the validated timesheets. However he can see the
total of effectives hours of all timesheets in the task and not only the
ones he can see.

This commit improves the filter by updated the
`hr_timesheet.timesheet_line_rule_portal_user` rule to correctly filter
the timesheets based on the setting instead of filtering the timesheets
manually in the read method.
Moreover, the `effective_hours`, `remaining_hours`, `total_hours_spent` and
`progress` has been removed to portal fields to compute the correct values
based on the timesheets the current user can see and not based to all
timesheets linked to the current tasks and the all subtasks of that task.

task-2648955

Community PR: odoo/odoo#82379

Closes #23315