[FIX] industry_fsm_sale: count quotations created via "Create Quotation"
Before this commit, the 'x Quotations' stat button in task form shown
all the quotations related to the task and not only the one created via
the `Create Quotation` button in the task form view.

This commit fixes this issue to correctly count the quotations added
into the task via the `Create Quotation` button.

task-2648955

Community PR: odoo/odoo#82379

Closes #23315