[FIX] industry_fsm{sale,stock}: prevent *_sale tests to use *_stock code:
industry_fsm_stock bring a new business flow where the SO is confirmed as soon as
a product is added to the fsm task in order to use the current user's default warehouse.

Although this behavior is needed in industry_fsm_stock we do not see any reason to force
it prior to stock installation. As such, the industry_fsm_sale tests are written not
taking this into account and fails if they are run post-install.

This commit also makes the necessary changes in order not to couple industry_fsm_sale and industry_fsm_stock tests by extracting the setUpClass.

task-2714003
Closes #22899

X-original-commit: fdd0a09
Part-of: odoo/enterprise#23373