[FIX] appointment: fix js test data
With the dynamic data based on the date of execution of the test,
it's possible that the time set will overlap with the previous or
next week. To avoid this, we set the date of the event as the 5th
January of the next year between 10 and 11 AM and also one event
a week after.

Indeed, for example the end of the week is a saturday and the start
of the week a sunday. Then if the test is executed during the night
at 1 AM, it was possible that the events were located in previous
week. This was leading to an error as the test couldn't find the
event in the view.

closes odoo/enterprise#23099

X-original-commit: 4ec9cc1c4e0235ae06e4941e0edeb043d67c45e7
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>