[FIX] stock_barcode: consider UoM when scanning weight
When processing a transfer, if the user scans a weight, it may lead to
incorrect result.

To reproduce the issue:
1. In Settings, enable "Units of Measure"
2. In Units of Measure Categories > Weight, add a new line unit of
measure:
    - Name: dag (decagram)
    - Type: Smaller
    - Ratio: 100
3. Create a product P:
    - UoM: g
    - Barcode: 2100123000005
4. In Inventory, create a planned receipt R:
    - 1 dag x P
5. Mark R as todo
6. Open Barcode > Operations > Receipts > R
7. Scan 10g of P (barcode: 2100123100002)

Error: The line is completed (1 / 1 dag), but another line is created
with a quantity of 9 dag

There are two errors during the process. The first one is in
`_processBarcode`: at some point, the exceeding quantity is computed:
https://github.com/odoo/enterprise/blob/cc539f0e6e8a82e02abe26e1d2ad4e3285244938/stock_barcode/static/src/models/barcode_model.js#L1196-L1202
But the calculations do not consider the UoMs. When scanning a weight,
the latter is expressed in default product's UoM (g in the above case).
However, the SM's UoM may be different (dag in above case). Therefore,
using `barcodeData.quantity` and `remainingQty` in the same calculations
does not make sense.

The second issue was actually hidden by the first one. Suppose the
exceeding quantity is correctly converted in `_processBarcode`, when
trying again the above case, it will lead to a new issue: the line will
be wrongly completed: 10 / 1 dag. This is due to `_updateLineQty`: the
new quantity is converted to the product's UoM which is wrong since
`qty_done` is based on the SML's UoM.

Linked to OPW-2705898

X-original-commit: cbc25b191f40e763fe640f47f9d800a745930e59
Part-of: odoo/enterprise#23268