[IMP] industry_fsm_sale: allow portal user to see invoices of task
Before this commit, when the portal user is in the project sharing
feature, he can see the invoices stat button in the task form view.
However, this button is disabled because we cannot redirect the user to
a portal view to see only the invoices related to this task.

This commit adds new routes to show the invoices related to a task. With
this route we can allow the portal user to click on the 'Invoices' stat
button in a form view of task when he is in the project sharing feature.
Moreover, a new field called `portal_invoice_count` field is added in
`project.task` model to count only the invoice that the portal user can
see. So this field will no longer be a related field of
`sale_order_id.invoice_count`. Instead, a compute method is defined for
this field.

task-2648955

Community PR: odoo/odoo#82379

Closes #23315