[IMP] delivery_ups: add commercial invoice
To generate a commercial invoice, UPS needed more information about the
commodities.

This commit does not change the code structure, it only adds the set_invoice
function as well as the Commodity class and the _get_commodities
function.

I also added the 'is_return' boolean in the shipment info to be able to
modify the ReasonForExport from SALE to RETURN.

I also added the 'invoice_required' boolean in the shipment info to only
create invoices in case of  international shipping.

The `factory_ns4` field is only used with the Ship api, and the Rate API
only defines 3 namespaces, which means it would crash if we try to retrieve
this 4th namespace.

Part-of: odoo/enterprise#22957