[ADD] l10n_mx_edi_stock: enable CartaPorte XML signing by PAC
The MX government requires all road deliveries to have a CartaPorte XML document sent to the PAC for signing.
This module enables the feature to generate and send a CFDI XML for CartaPorte, and the ability to print these details on the Delivery Slip PDF.

Since the XML document also requires the distance in KM, this module also has a distance calculator that uses the MapBox API to compute the distance from source address to destination address.

Task-2585661

X-original-commit: 92b2ab9d02d2b4f891396ed12d72786baa789679
Part-of: odoo/enterprise#23244