[ADD] l10n_mx_edi_stock_extended: COMEX features
Mexico exports (AKA comex) provides the ability to generate, sign and print the delivery guide for international road deliveries. Since not all MX users require this feature, we have split it into a separate module that depends on l10n_mx_edi_extended. In this way, we are able to keep the UI clean and simple (without the extended fields) for most of our users.

Task-2585661

closes odoo/enterprise#23244

X-original-commit: 9f4809bf7472a6d077a20af146d44df782ddfb3a
Related: odoo/odoo#82224
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Ayob Habib (ayh) <ayh@odoo.com>