[IMP] industry_fsm_sale: rename the action used in xpath
This commit renames name attribute of the button used in a xpath because
this button is in fact the stat button to show the parent task in the
project sharing feature. This action has been renamed into
`action_project_sharing_view_parent_task` in project module.

task-2648955

Community PR: odoo/odoo#82379

Closes #23315