[IMP] delivery_fedex: use country_of_origin for CI
Before this commit, the value for commodity_country_of_manufacture was
given by the warehouse country, but it makes more sense to use a
special field for it, as explained in this commit:
odoo/odoo@7fc9234

In this file, the `commodity_country_of_manufacture` was placed right
after the `commodity_description` field, because it is the same order as
the arguments in the `srm.commodities()` function.

To avoid sending blank fields - which would cause the shipping to fail -
if the country_of_origin field is not filled, the default behaviour is
the same as before this commit: it then takes the warehouse country.

task-2701428

closes odoo/enterprise#22903

Signed-off-by: Arnold Moyaux <arm@odoo.com>