[IMP] delivery_easypost: CI origin_of_goods + lognote
Before this commit, the value for custom_infos->origin_country was
given by the warehouse country, but it makes more sense to use a
special field for it, as explained in this commit:
odoo/odoo@7fc9234

To avoid sending blank fields - which would cause the shipping to fail -
if the country_of_origin field is not filled, the default behaviour is
the same as before this commit: it then takes the warehouse country.

Also, we can now retrieve the commercial invoice and any documents that
would be sent by Easypost in the picking log notes.

When retrieving the documents from the furnished link, there was no
timeout for the get request so I added one, and in case it cannot be
downloaded in time (30s), a link is given in the log note instead.

Part-of: odoo/enterprise#22907