[FIX] l10n_pe_edi: detractions need to be discounted from the payments due
Withholding or detractions in PE are handled as a payment.  And that is why
we need to add them in the payment terms.

On the one hand, we need to add it as a separate section detraction in the
company currency.  On the other hand, we need to subtract it from the first
payment term line in the currency of the invoice.

We added a test with the combination of foreign currency, enhanced payment
term and detraction.

Original PR from Luigys from Vauxoo: #22573

closes odoo/enterprise#23127

X-original-commit: 830145196b385e4665a71fce6b8a6f87fda0a6d2
Signed-off-by: Laurent Smet <las@odoo.com>