[IMP] stock_barcode_picking_batch: rearrange package form view
The addition of `package_use` to the package form view was done in a way
where it isn't possible to add extra fields to the 2nd column of the
form in the stock module. Therefore we update it so we can add a new
field to the 2nd column and have this field after that.

Part of Task: 2494740
Related COM PR: odoo/odoo#82389

Part-of: odoo/enterprise#23323