[FIX] mrp_mps: prevent line from disappearing
When adding a new product, if the user then clicks on "Replenish", the
product disappears

To reproduce the issue:
(Use demo data)
1. In MPS, add a product: [FURN_78236] Table Kit
2. Click on "Replenish" for this specific product

Error: The line disappears. When loading again the MPS page, the line is
back and clicking again on "Replenish" won't reproduce the issue

Once the replenishment is done, the page is reloaded:
https://github.com/odoo/enterprise/blob/b7168a46901aacbb4957fa17cc44a70ba2ce1cf6/mrp_mps/static/src/js/client_action.js#L185-L192
and this will lead to
https://github.com/odoo/enterprise/blob/b7168a46901aacbb4957fa17cc44a70ba2ce1cf6/mrp_mps/static/src/js/client_action.js#L284-L291
But here is the issue: when adding the product (step 01), nothing adds
the new line in `active_ids`. This is the reason why this line
"disappears"

OPW-2662244

closes odoo/enterprise#23073

X-original-commit: d521a6c604714af1ffc745cd2074a9e161664ce8
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>