[FIX] account_followup: fix partner.total_due compute
Followup messages can contain variables. The "%(amount_due)s" variable is replaced with the "partner.total_due" value.
Before this fix, partner.total_due was the sum of all unpaid invoices, blocked or not. This led to an inconsistency between the "Total Due" amount in the followup report (that would not include blocked invoices) and the amount in the followup message (including blocked invoices).
With this fix, blocked invoices are not added to "partner.total_due" anymore.
The followup report view is also updated now when total_due value changes.

closes odoo/enterprise#23049

Task: 2684627
X-original-commit: 3d3dcce167f135514111762818477bc065b7d1b6
Signed-off-by: Olivier Colson <oco@odoo.com>