[IMP] web_grid: add read_grid in batch mode
This commit implements the read_grid_grouped which returns the web_grid data
at once, instead of per section (group) as it is when calling read_grid.

The current implementation is too slow when called on big database.

Thanks to this commit, the number of calls to the database to fetch de grid data
is lowered to 2 instead of 1 + #groups.

task-2700429
Related PR: odoo/odoo#81894

closes odoo/enterprise#23100

X-original-commit: 5cb5698739003f4d68fc082a619b1e6f15c1ae67
Related: odoo/odoo#81907
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Co-authored-by: Xavier BOL (xbo) <xbo@odoo.com>
Co-authored-by: Priyanka Kaakdiya <pka@odoo.com>
Co-authored-by: Mohammed Shekha <msh@odoo.com>