[IMP] sign: support copy sign request
allow users to `copy` a sign request.
Copy a sign request means using the configure of the current sign request to send a new sign request
So, the state of the new sign request and its sign request items are `sent`.
It can be used to resend a canceled/refused sign request

this commit also separates the attachments of the sign request and the attachments for the completed documents

Before this commit:
sign_reuqest.copy() is not allowed and checked

Taskid: 2659342
Part-of: odoo/enterprise#22813