[FIX] account_followup: mail sent not correct using cron
When the cron to send the followup report ("Account Report Followup; Execute followup") is triggered,
the email with the followup report sent might not reflect the followup level set in the database.
For example, if you rely only on automatic followup without checking the follow-up report in the interface,
the system will send emails only with the message of the 1st followup level.

The option 'keep_summary' is set by default when we call the send_email() function. This will keep the
last summary set in the report manager. This is ok if we use the interface to send mail because the summary
is updated to the one of the current followup level when loading the followup report page. In the case of
the schedule action (which uses the same function to send mail), this is not because it might sent an
outdated mail (with different amounts/invoices/dates).

This commit removes the default 'keep_summary' = True and set this parameter when we click on the button
"Send by mail" in the follower. So if the cron is triggered it will recompute the summary (mail message)
based on the followup level.

opw-2667996

closes odoo/enterprise#23254

X-original-commit: fdfb06359841627e580e3feffd98e3f088cba261
Signed-off-by: William André (wan) <wan@odoo.com>