[REM] project_*: remove some stat button in project update
This commit removes stat buttons displayed in the right side panel in
project update kanban view to prepare the revamp of the content shown in
that project update. Indeed, the content in the right side panel will
show the information shown in those stat buttons but with more details.

task-2710808