[FIX] l10n_ae_hr_payroll_account: check accounts before setup
Currently when this modules is installed/udpated we setup the accounting
info using the new account codes for the EA's CoA in 15.0
https://github.com/odoo/odoo/commit/8d6bd85d2ba9f1aee414734372c0ea684cbd6aa5

Before this patch we used to fail the setup with a message to the user.
Here we silently skip the setup if the necessary accounts are not found.

This is far from ideal, and the user will need to manually setup the
account codes. It is, nonetheless, necessary to allow the upgrade of
older DBs with the AE CoA already installed.

closes odoo/enterprise#23224

X-original-commit: e8e9018a9802858374cf6b119768f67c1f1854a7
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>