[FIX]sale_project_forecast: fix so_line compute in planning slot
Purpose of this commit to set sale order line field default
for slot depending sale order line set on task or project
of that slot.

So, in this commit remove extra code that stop compute of
sale_line_id to get value from task or project.

task-2732325

Closes:  #23374
X-original-commit: cb7813168d3628cc528d5ce7e81eabf03ddbe418
Part-of: odoo/enterprise#24810