[FIX] social_twitter: fix a tour that randomly failed
Bug
===
The tour to check the limit of replies for social Twitter randomly fail.

There are 2 reason for that
1. Even if the view all Tweets were below the right message, the mock
   of the API returned the wrong value for "in_reply_to_status_id_str"
2. The check of the last message is done by checking the message in the
   DOM, but as it's fast enough, we didn't catch it.

So the test fail when we correctly detect that the message is posted
(and it shouldn't).

So to fix this issue, we return the right value for "in_reply_to_status_id_str",
and we add an "extra_trigger" to the tour to wait before checking if the
message is posted.

Just in case, remove randomly generated Twitter identifiers and use a
counter instead.

closes odoo/enterprise#23135

X-original-commit: e2e93414d89301764a3ec70699bbefea2d32dc5e
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>