[FIX] project_account_budget: fix stat button in right side panel
Purpose of this commit is to only show 'Budget' stat button
in right side panel of project update when analytic account set
on that project.

So, this commit adds a condition in stat button to show that one
only when analytic account is set on that project.

task-2722863

closes odoo/enterprise#23445

Closes: https://github.com/odoo/enterprise/pull/23179
X-original-commit: e24d71677edb39366bfe6dceb5094a28ac9be635
Related: odoo/odoo#82649
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>