[FIX] l10n_cl_edi_stock: if there is no activity description in partner or commercial partner raise an error. Also, take the commercial partner activity and dte email if not present in the current partner
This also solves the problem of the availability of the document by a physical person. If the email is sent only to the reception dte_email, sometimes persons don't see the document in order to be aware promptly if the document has been sent. Thats why we include the partner_id.email to the send.
Also, the SII resolution says that the Contacto is a person email

closes odoo/enterprise#25025

X-original-commit: b58fddf511c10a6aa750dd20621835a6428d4026
Signed-off-by: Josse Colpaert <jco@odoo.com>