[FIX] l10n_pe_edi: When sending invoice complement via IAP, retrieve the CDR if the sending fails
with the error that the invoice was already registered.

The issue:
Sometimes, an invoice can have been correctly registered with SUNAT but Odoo may not have correctly
understood this. This can happen, for example, if the cron which sends the payment complement
crashes because of a concurrent access error.

When this happens, the Odoo user does not have access to the CDR (the receipt from SUNAT). This
is an issue because they need to send it to their customer.

If the user retries sending the invoice, the IAP service returns an error because the invoice is
already registered. It does not return the CDR, either.

As a result, the user has no way of obtaining the CDR.
Furthermore, the state in Odoo is not aligned with the state in SUNAT, and the user is not
presented with the relevant buttons e.g. to request the invoice's cancellation.

The solution:
When sending the invoice via the IAP service, if the IAP service returns an error because the
invoice is already registered, re-request the CDR from the IAP service and advance the state
in Odoo accordingly.

We modify the error-handling logic when sending the invoice via the IAP service:
- First, check whether the error code is 1033, which means that the invoice was rejected
  because it was already registered with the OSE.
- If the error code is 1033:
    - Try to retrieve the CDR status via the new method _l10n_pe_edi_retrieve_cdr_iap.
    - If retrieving the CDR status also fails, return an error.
    - If retrieving the CDR status yields a CDR status code which is different from '0004'
      (the CDR already exists), also return an error.
    - If retrieving the CDR status yields a CDR status code of '0004', extract the CDR from
      the response and advance the state in Odoo as if the invoice was successfully sent.
- If the error code is different from 1033, raise an error as usual.

Related to support tickets 2714559 and 2663502.

closes odoo/enterprise#23570

X-original-commit: 68663e0cab59c144f52421e92276e74e2da5341d
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>