[IMP] delivery_ups: add shipping insurance
Shipping insurance is a service which may reimburse senders whose parcels are
lost, stolen, and/or damaged in transit. With UPS, we can declare a value
for each package. This value corresponds to the cost of the products in
the package.

Since the currency could vary from one picking to another, the currency
in which the declared values are sent is the vurrency of the company
linked to the order (rate) or to the picking (ship).

Because we do not know what is in each package at the time of rating,
the total cost is distributed equally between the packages.

The `picking.move_line_ids_without_package` field does not seem to work
properly, so I used a filter to get the stock_move_lines without any
package_id.

There are 2 types of code for DeclaredValue.Type.Code: 01 (EVS), 02
(DVS). The 01 can be used in more cases.

PS:This whole code could potentially go to the base module for delivery.

task-2701459
PR in base: https://github.com/odoo/odoo/pull/83896

Part-of: odoo/enterprise#22957