[FIX] marketing_automation: Wrong default campaign_id in context
When sending the email from marketing.participant, the wrong default campaign_id (set in action marketing_participant_action_campaign)
was used to create a mail compose message. Odoo expects the utm_campaign_id to track
the links and not the id of the campaign. It lead to issues if the id of the campaign doesn't exist in the table utm.campaign

PS: On model mail.compose.message, the related field to utm.campaign is called campaign_id

opw:2555207

closes odoo/enterprise#23174

X-original-commit: f9b4f353c8c0422bcaba07be836856b2f6776953
Signed-off-by: Simon Goffin <sig@odoo.com>