[FIX] l10n_be_hr_payroll: allow changing working schedule
Currently, if an employee doesn't have any holiday allocation, the working schedule wizard
doesn't have any selection for the time off type as it requires to have a valid allocation.
As this field is mandatory, it is not possible to change the working schedule if the employee
doesn't have an allocation.

The working schedule wizard should list all time off type that requires an allocation to prevent
this case.

task-2701128

closes odoo/enterprise#25843

X-original-commit: 0736a3e890313c07bad5bf245a820651b4a11932
Related: odoo/odoo#87845
Signed-off-by: Kevin Baptiste <kba@odoo.com>