[FIX] l10n_mx_edi: round exch rate up in payment complement
The issue:
When validating the payment CFDI, the Mexican government checks that the
`Monto` amount of the payment is greater or equal to the sum of the
`ImpPagado / TipoCambioDR` amounts for the invoices.

This check sometimes fails for large invoices in USD where the payment is in
MXN, because the TipoCambioDR field is rounded to six decimals.

For example, take the following situation:
- An invoice was issued for USD 20,000
- When it was paid, the rate of the day was 0.0500004 USD/MXN
- So the customer paid MXN 399996,80
- In the CFDI, the TipoCambioDR is rounded down to 0.050000 USD/MXN
- So ImpPagado / TipoCambioDR = 20,000 / 0.05 = 400000 USD
- This is greater than the 399996,80 of the payment, so the check fails.

The solution:
Round the TipoCambioDR up, rather than down.

Note:

closes odoo/enterprise#23317

Monto: the total amount paid by the payment (e.g. in MXN)
Imppagado: the portion of the invoice paid by the payment (e.g. in USD)
Tipocambiodr: the exchange rate between the payment and the invoice
X-original-commit: 7c43ba6b5b5306dfe49725ac0161566bde9cf7b3
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>